Add NaivePredictor

+++ src/main/java/ru/msu/cs/svdtop/domain/Predictor.java
@@ -6,5 +6,5 @@ import ru.yandex.bolts.collection.ListF;
  * @author lifar
  */
 public interface Predictor<TSnapshot extends Snapshot> {
-    ListF<Long> getTop(TSnapshot snapshot, Profile userProfile, long count);
+    ListF<Long> getTop(TSnapshot snapshot, Profile userProfile, int count);
 }
